test(avatar): recategorize and clean up browser tests
